Add files via upload
This is our first attempt of the Use Case Diagram for the Call Management Center (CMC). We decided to create two Use Case diagrams to represent the inbound and outbound process occurring from the Customer end, and Relationship  Manager end, respectively. We also decided not to include the profile matching system as an Actor because it was not a stakeholder. However we will address this process further in our class diagram. We still will review and update this current Use Case Diagram